requirements.txt
Flask==3.0.3
python-docx==1.1.2
Werkzeug==3.0.3
gunicorn==21.2.0
streamlit
